+++ index.tsx
@@ -19,7 +19,7 @@ if (!rootElement) throw new Error("Could not find root element to mount to");
 const root = ReactDOM.createRoot(rootElement);
 root.render(
   <React.StrictMode>
-    <ClerkProvider publishableKey={CLERK_PUBLISHABLE_KEY}>
+    <ClerkProvider publishableKey={CLERK_PUBLISHABLE_KEY} clerkJSVersion="6.30.1">
       <App />
     </ClerkProvider>
   </React.StrictMode>